upgrade server and client: server can get messages from client for all time
